python/preview/declarative_self_improvement/dspy_fewshot/requirements.txt
dspy==3.1.3
ollama==0.6.1
openai==2.21.0
pandas==3.0.0
pydantic==2.12.5
python-dotenv==1.2.1
python_dateutil==2.9.0.post0
Requests==2.32.5
